requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
docxtpl
pgvector
sentence-transformers
numpy
faiss-cpu
torch
passlib
ollama
pdfplumber
pythainlp
python-multipart